Move middleware stack to active_support
